Enable open source multi akm for specific Broadcom project
For cross akm roaming function, we have two ways in supplicant :
1. Open source way.
2. Broadcom specific way.

Originally, Broadcom projects follow way 2 due to the limitation of the driver.
For now, we plan to move all Broadcom projects to follow way 1 to make supplicant code clean and downsteam without conflict.

For those projects which is ready for way 1, we control it to follow way 1 with WIFI_BRCM_OPEN_SOURCE_MULTI_AKM config.
For the rest of Broadcom projects, we still follow way 2 until the multiple AKMs is fully supported by driver and verified.

For ongoing Broadcom project, we also need to exercise the open source path in supplicant to do multiple AKM configuration to comply with WFA Wifi 7 AKM24 specification (b/331896542 ).

Bug: 246871098
Test: AU drop test
Test: Cross AKM roaming test
Flag: EXEMPT enable control flag for aosp akms
Change-Id: I57fadcc164b23d6bdb85a0909d23dfa32fc71b12